+++ src/onegov.directory/onegov/directory/errors.py
@@ -13,3 +13,9 @@ class MissingColumnError(OnegovDirectoryError):
     def __init__(self, column, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.column = column
+
+
+class DuplicateEntryError(OnegovDirectoryError):
+    def __init__(self, name, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.name = name